Comments and reorganization
